# backend-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.rateLimitRuleRequest:
    properties:
      blockDuration:
        maximum: 604800
        minimum: 0
        type: integer
      burst:
        type: integer
      cooldown:
        maximum: 86400
        minimum: 0
        type: integer
      description:
        maxLength: 500
        type: string
      enabled:
        type: boolean
      maxViolations:
        maximum: 10000
        minimum: 0
        type: integer
      method:
        type: string
      pathPattern:
        type: string
      priority:
        maximum: 9999
        minimum: 0
        type: integer
      rate:
        type: number
      violationScore:
        maximum: 1000
        minimum: 0
        type: integer
    required:
    - burst
    - pathPattern
    - rate
    type: object
...
  model.RateLimitRule:
    properties:
      blockDuration:
        description: 封禁时长（秒），超过触发次数后封禁 IP
        type: integer
      burst:
        description: 突发容量
        type: integer
      cooldown:
        description: 冷却时间（秒），触发限流后需等待多久恢复
        type: integer
      createdAt:
        type: string
      description:
        description: 规则描述
        type: string
      enabled:
        description: 是否启用
        type: boolean
      id:
        type: integer
      maxViolations:
        description: 最大违规次数，超过后触发封禁（0=不封禁）
        type: integer
      method:
        description: HTTP 方法：GET/POST/PUT/DELETE/* (所有)
        type: string
      pathPattern:
        description: 路径模式，支持 * 通配符
        type: string
      priority:
        description: 优先级，数值越大越先匹配
        type: integer
      rate:
        description: 每秒请求数
        type: number
      updatedAt:
        type: string
      violationScore:
        description: 违规风险分，触发限流时累加到风险评分系统（0=不累加）
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
  service.LoginResponse:
    properties:
      accessToken:
        type: string
      avatar:
        type: string
      bio:
        type: string
      birthday:
        type: string
      email:
        type: string
      gender:
        type: integer
      homePath:
        type: string
      id:
        type: integer
      mustChangePassword:
        description: 首次登录或管理员重置密码后需强制修改
        type: boolean
      nickname:
        type: string
      phone:
        type: string
      realName:
        type: string
      refreshToken:
        type: string
      registerSource:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
...
  service.OAuthBindingItem:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      provider:
        type: string
      providerAvatar:
        type: string
      providerUsername:
        type: string
    type: object
  service.OAuthUnbindRequest:
    properties:
      provider:
        type: string
    required:
    - provider
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 后端管理服务 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Server API
  version: "1.0"
paths:
...
  /auth/oauth/authorize:
    get:
      description: 获取第三方平台的授权页面 URL
      parameters:
      - description: 提供商：wechat/github/google
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取第三方授权 URL
      tags:
      - OAuth绑定
  /auth/oauth/bind-url:
    get:
      description: 获取第三方平台的授权页面 URL
      parameters:
      - description: 提供商：wechat/github/google
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取第三方授权 URL
      tags:
      - OAuth绑定
  /auth/oauth/bindings:
    get:
      description: 获取当前用户已绑定的第三方账号列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.OAuthBindingItem'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取第三方绑定列表
      tags:
      - OAuth绑定
  /auth/oauth/callback:
    get:
      description: 第三方平台授权完成后的回调接口，返回 JWT token
      parameters:
      - description: 提供商
        in: query
        name: provider
        required: true
        type: string
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态参数
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 第三方授权回调
      tags:
      - OAuth绑定
  /auth/oauth/unbind:
    post:
      consumes:
      - application/json
      description: 解绑指定的第三方账号
      parameters:
      - description: 解绑请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.OAuthUnbindRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 解绑第三方账号
      tags:
      - OAuth绑定
...
  /system/rate-limit-rules:
    get:
      description: 获取所有接口限流规则
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.RateLimitRule'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取限流规则列表
      tags:
      - 限流规则
    post:
      consumes:
      - application/json
      description: 创建新的接口限流规则
      parameters:
      - description: 限流规则
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.rateLimitRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RateLimitRule'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建限流规则
      tags:
      - 限流规则
  /system/rate-limit-rules/{id}:
    delete:
      description: 删除指定接口限流规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除限流规则
      tags:
      - 限流规则
    get:
      description: 根据 ID 获取限流规则详情
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RateLimitRule'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取限流规则详情
      tags:
      - 限流规则
    put:
      consumes:
      - application/json
      description: 更新指定接口限流规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 限流规则
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.rateLimitRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RateLimitRule'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新限流规则
      tags:
      - 限流规则
  /system/rate-limit-rules/{id}/status:
    put:
      consumes:
      - application/json
      description: 启用或禁用指定限流规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 启用状态
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新限流规则状态
      tags:
      - 限流规则
...
securityDefinitions:
  BearerAuth:
    description: JWT Token（格式：Bearer {token}）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"